+++ release-notes.md
@@ -16,6 +16,7 @@ See [log](https://github.com/Pathoschild/SMAPI/compare/1.9...1.10).
 For players:
 * Added support for Stardew Valley 1.2 beta.
 * Added logic to rewrite many mods for compatibility with game updates, though some mods may still need an update.
+* Added `world_setyear` console command to TrainerMod.
 * Fixed some players getting `SEHException` errors.
 * Fixed rare issue where the installer would crash trying to delete a bundled mod from `%appdata%`.